+++ lib/main.dart
@@ -1,9 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
+
 import 'package:w3/ui/screens/login/login_screen.dart';
+import 'package:provider/provider.dart';
+import 'package:w3/ui/screens/product_details/product_details.dart';
 
 void main() {
-  runApp(const MyApp());
+  runApp(
+    MultiProvider(providers: [
+      Provider(create: (context) => ProductDeatailsScreen()),
+    ], child: MyApp()),
+  );
 }
 
 class MyApp extends StatelessWidget {
