+++ lib/ui/screens/product_details/product_details.dart
@@ -1,0 +1,150 @@
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:w3/core/constants/colors_constants.dart';
+
+import 'package:w3/ui/custom_widgets/custom_widgets.dart';
+import 'package:w3/core/constants/text_style.dart';
+import 'package:bubble_tab_indicator/bubble_tab_indicator.dart';
+
+import 'package:w3/core/constants/buttons_style.dart';
+import 'package:w3/ui/screens/signup/signup_screen.dart';
+
+class ProductDeatailsScreen extends StatefulWidget {
+  const ProductDeatailsScreen({Key? key}) : super(key: key);
+
+  @override
+  State<ProductDeatailsScreen> createState() => _ProductDeatailsScreenState();
+}
+
+class _ProductDeatailsScreenState extends State<ProductDeatailsScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: kTextFieldFilledColor,
+      body: Stack(children: [
+        Positioned(
+          top: 40.h,
+          left: 30.w,
+          right: 30.w,
+          child: Image(
+            height: 450.h,
+            width: 250.w,
+            image: AssetImage('assets/images/plant1.png'),
+          ),
+        ),
+        DraggableScrollableSheet(
+          initialChildSize: 0.3.h,
+          // minChildSize: 0.30,
+          //maxChildSize: 0.9,
+          builder: (BuildContext context, ScrollController scrollController) {
+            return Container(
+              child: Padding(
+                padding: EdgeInsets.only(
+                    top: 30.h, left: 20.w, right: 20.w, bottom: 10.h),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: [
+                        Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: [
+                            Text(
+                              'Schefflera',
+                              style: kH1TextStyle.copyWith(fontSize: 20.sp),
+                            ),
+                            Row(
+                              children: [
+                                Text("\$25.00",
+                                    style: kH2TextStyle.copyWith(
+                                        color: kGreen,
+                                        fontWeight: FontWeight.bold)),
+                                SizedBox(
+                                  width: 5.w,
+                                ),
+                                Icon(
+                                  Icons.star,
+                                  color: Colors.yellow,
+                                  size: 15.h,
+                                ),
+                              ],
+                            ),
+                          ],
+                        ),
+                        Column(
+                          children: [
+                            Container(
+                              decoration: BoxDecoration(
+                                color: kDarkGreen,
+                                borderRadius: BorderRadius.circular(20.r),
+                              ),
+                              height: 25.h,
+                              width: 90.w,
+                              child: Center(
+                                child: Row(
+                                  children: [
+                                    Expanded(
+                                      child: IconButton(
+                                          onPressed: null,
+                                          icon: Icon(
+                                            Icons.remove,
+                                            color: Colors.white,
+                                            size: 12.h,
+                                          )),
+                                    ),
+                                    const Expanded(
+                                        child: Center(
+                                            child: Text(
+                                      '10',
+                                      style: TextStyle(color: Colors.white),
+                                    ))),
+                                    Expanded(
+                                      child: IconButton(
+                                          onPressed: null,
+                                          icon: Icon(
+                                            Icons.add,
+                                            color: Colors.white,
+                                            size: 12.h,
+                                          )),
+                                    ),
+                                  ],
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                      ],
+                    ),
+                    SizedBox(
+                      height: 20.h,
+                    ),
+                    Text("About",
+                        style: kH1TextStyle.copyWith(fontSize: 14.sp)),
+                    Text(
+                      "However, they look like huge white flowers and they bloom throughout the year and a bit more frequently in the springtime. This coupied with plant's brood, deep green leaves.",
+                      style: TextStyle(
+                        fontSize: 10.sp,
+                        height: 1.5.h,
+                        color: kDarkGreen,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+
+              // container decoration
+              decoration: BoxDecoration(
+                  color: Colors.white,
+                  borderRadius: BorderRadius.only(
+                      topLeft: Radius.circular(40.r),
+                      topRight: Radius.circular(40.r))),
+            );
+          },
+        ),
+      ]),
+    );
+  }
+}
