+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   flutter_screenutil: ^5.4.0+1
   bubble_tab_indicator: "^0.1.5"
+  provider: ^6.0.2
 
 dev_dependencies:
   flutter_test:
